ocr/requirements.txt
# OCR CLI Dependencies
pytesseract>=0.3.10
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0